Move upload logic to `ECGModule`
