+++ PAWS/ECGRecordings/ECGModule.swift
@@ -6,9 +6,15 @@
 // SPDX-License-Identifier: MIT
 //
 
+import FirebaseAuth
+import FirebaseFirestore
 import HealthKit
+import OSLog
+import enum ModelsR4.ResourceProxy
 import Spezi
 import SpeziLocalStorage
+import SpeziMockWebService
+import UserNotifications
 
 
 @Observable
@@ -20,9 +26,12 @@ class ECGModule: Module, DefaultInitializable, EnvironmentAccessible {
     
     @ObservationIgnored @Dependency var localStorage: LocalStorage
     @ObservationIgnored @StandardActor var standard: PAWSStandard
+    @ObservationIgnored @Dependency var mockWebService: MockWebService?
     
     private(set) var electrocardiograms: [HKElectrocardiogram] = []
     private var uploadedElectrocardiograms: Set<HKElectrocardiogram.ID> = []
+    private let healthStore = HKHealthStore()
+    private let logger = Logger(subsystem: "PAWS", category: "Standard")
     
     
     /// Creates an instance of a ``MockWebService``.
@@ -95,5 +104,114 @@ class ECGModule: Module, DefaultInitializable, EnvironmentAccessible {
         }
         
         return nil
+    }
+    
+    func updateElectrocardiogram(basedOn categorySample: HKCategorySample) async {
+        do {
+            guard let updatedElectrocardiogram = try await self.electrocardiogram(
+                correlatedWith: categorySample,
+                from: healthStore
+            ) else {
+                return
+            }
+            
+            try await upload(sample: updatedElectrocardiogram, force: true)
+        } catch {
+            logger.log("Could not corrolate category sample with ECG: \(categorySample)")
+        }
+    }
+    
+    private func upload(sample: HKSample, force: Bool = false) async throws {
+        let resource: ResourceProxy
+        if let electrocardiogram = sample as? HKElectrocardiogram {
+            self.insert(electrocardiogram: electrocardiogram)
+            
+            guard !self.isUploaded(electrocardiogram) || force else {
+                return
+            }
+            
+            async let symptoms = try electrocardiogram.symptoms(from: healthStore)
+            async let voltageMeasurements = try electrocardiogram.voltageMeasurements(from: healthStore)
+            
+            resource = ResourceProxy(
+                with: try await electrocardiogram.observation(
+                    symptoms: symptoms,
+                    voltageMeasurements: voltageMeasurements
+                )
+            )
+        } else {
+            resource = try sample.resource
+        }
+        
+        if let mockWebService {
+            let encoder = JSONEncoder()
+            encoder.outputFormatting = [.prettyPrinted, .sortedKeys, .withoutEscapingSlashes]
+            let jsonRepresentation = (try? String(data: encoder.encode(resource), encoding: .utf8)) ?? ""
+            try await mockWebService.upload(path: "healthkit/\(sample.uuid.uuidString)", body: jsonRepresentation)
+        } else {
+            try await standard.healthKitDocument(id: sample.id).setData(from: resource)
+        }
+        
+        if let electrocardiogram = sample as? HKElectrocardiogram {
+            self.markAsUploaded(electrocardiogram)
+        }
+    }
+    
+    func upload(electrocardiogram: HKElectrocardiogram) async {
+        var supplementalMetrics: [HKSample] = []
+        
+        do {
+            try await upload(sample: electrocardiogram)
+            
+            supplementalMetrics.append(contentsOf: (try? await electrocardiogram.precedingPulseRates) ?? [])
+            supplementalMetrics.append(contentsOf: (try? await electrocardiogram.precedingPhysicalEffort) ?? [])
+            supplementalMetrics.append(contentsOf: (try? await electrocardiogram.precedingStepCount) ?? [])
+            supplementalMetrics.append(contentsOf: (try? await electrocardiogram.precedingActiveEnergy) ?? [])
+            
+            if let precedingVo2Max = try? await electrocardiogram.precedingVo2Max {
+                supplementalMetrics.append(precedingVo2Max)
+            }
+            
+            for supplementalMetric in supplementalMetrics {
+                do {
+                    try await upload(sample: supplementalMetric)
+                } catch {
+                    logger.log("Could not upload \(supplementalMetric.sampleType): \(error)")
+                    let content = UNMutableNotificationContent()
+                    content.title = "Upload Error"
+                    content.body = "Sample could not be uploaded \(supplementalMetric.sampleType.description) (\(supplementalMetric.uuid.uuidString) at \(Date.now.formatted(date: .numeric, time: .complete)): \((supplementalMetric as? HKQuantitySample)?.quantity.description ?? "Unknown")"
+                    let request = UNNotificationRequest(identifier: UUID().uuidString, content: content, trigger: nil)
+                    try? await UNUserNotificationCenter.current().add(request)
+                }
+            }
+        } catch {
+            logger.log("Could not access HealthKit sample: \(error)")
+            let content = UNMutableNotificationContent()
+            content.title = "HealthKit Error"
+            content.body = "Sample \(electrocardiogram.sampleType.description) with identifier \(electrocardiogram.uuid.uuidString)"
+            let request = UNNotificationRequest(identifier: UUID().uuidString, content: content, trigger: nil)
+            try? await UNUserNotificationCenter.current().add(request)
+        }
+    }
+    
+    private func uploadUnuploadedECGs() {
+        for ecg in electrocardiograms where !isUploaded(ecg) {
+            Task {
+                do {
+                    try await self.upload(sample: ecg)
+                } catch {
+                    // Handle error
+                }
+            }
+        }
+    }
+}
+
+extension Array where Self.Element: Hashable {
+    mutating func appendUnique<C: Collection>(from source: C) where C.Element == Self.Element {
+        let existingItems = Set(self)
+        for item in source where !existingItems.contains(item) {
+            self.append(item)
+        }
     }
 }
